Factorise common filters code
